[URLParser] Fix file: as a relative file URL https://bugs.webkit.org/show_bug.cgi?id=169122
Reviewed by Tim Horton.

Source/WebCore:

This is clearly defined in https://url.spec.whatwg.org/#file-state with the EOF code point.
I got it wrong and didn't test it.  It's been added to the web platform tests since we last updated.

* platform/URLParser.cpp:
(WebCore::URLParser::parse):

Tools:

* TestWebKitAPI/Tests/WebCore/URLParser.cpp:
(TestWebKitAPI::TEST_F):

LayoutTests:

* TestExpectations:
We don't need to skip url-setters.html any more.  It used to assert before the new URLParser was enabled.
* fast/loader/url-parse-1-expected.txt:
* fast/url/file-expected.txt:
We used to add a strange and unnecessary :/// at the end of the URL.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213384 268f45cc-cd09-0410-ab3c-d52691b4dbfc