Add a script to synchronize FeatureDefines.xcconfig files https://bugs.webkit.org/show_bug.cgi?id=169151 <rdar://problem/30841856>
Reviewed by Simon Fraser.

I got sick of manually copying the FeatureDefines.xcconfig
files, so I added a script to do it. As long as you
run this after you've edited one, it will take the most
recently updated and copy it to all the others.

* Scripts/sync-feature-defines: Added.
* Scripts/webkitpy/style/checkers/featuredefines.py:
(FeatureDefinesChecker.check): Export the list of files,
and give a suggestion to use the sync script.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213395 268f45cc-cd09-0410-ab3c-d52691b4dbfc