LayoutTest fast/dom/timer-throttling-hidden-page.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=168927
Reviewed by Simon Fraser.

Source/WebCore:

Update internals.isTimerThrottled() to return true for any kind of throttling.
It used to only account only for specific timers that do not cause changes
inside the viewport. It now accounts for all kind of throttling, such as
timer alignment due to the page being hidden.

* page/DOMTimer.h:
* testing/Internals.cpp:
(WebCore::Internals::isTimerThrottled):

LayoutTests:

* fast/dom/nested-timer-visible-element-throttling.html:
* fast/dom/repeating-timer-element-overflow-throttling.html:
* fast/dom/repeating-timer-element-overflowing-child-own-layer-throttling.html:
* fast/dom/repeating-timer-mixed-visible-display-none-elements-throttling.html:
* fast/dom/repeating-timer-visible-element-throttling.html:
* fast/dom/timer-height-change-no-throttling.html:
* fast/frames/flattening/resources/iframe-to-resize.html:
Use 10 milliseconds interval for tests relying on internals.isTimerThrottled().
Using 0 is problematic because we always internally cap to 1ms and 4ms after
reaching the max nesting level. Therefore, a 0 delay timer is always considered
to be throttled.

* fast/dom/timer-throttling-hidden-page-expected.txt:
* fast/dom/timer-throttling-hidden-page-non-nested-expected.txt:
* fast/dom/timer-throttling-hidden-page-non-nested.html:
* fast/dom/timer-throttling-hidden-page.html:
Rewrite tests to stop measuring timer intervals as this is inherently
flaky and rely on internals.isTimerThrottled() instead.

* platform/mac/TestExpectations:
Unskip test that used to be flaky.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213458 268f45cc-cd09-0410-ab3c-d52691b4dbfc