[Modern Media Controls] Setting a valid source after an invalid one retains the invalid placard https://bugs.webkit.org/show_bug.cgi?id=169280 <rdar://problem/30876078>
Patch by Antoine Quint <graouts@apple.com> on 2017-03-07
Reviewed by Jon Lee.

Source/WebCore:

We need to track when a new load start to see if we might have recovered from an error
and hide the "invalid" placard.

Test: media/modern-media-controls/placard-support/placard-support-error-recover.html

* Modules/modern-media-controls/media/placard-support.js:
(PlacardSupport.prototype.get mediaEvents):

LayoutTests:

Add a new test that checks that we know to remove the "invalid" placard if a new source is
set and media starts loading again.

* media/modern-media-controls/placard-support/placard-support-error-recover-expected.txt: Added.
* media/modern-media-controls/placard-support/placard-support-error-recover.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213514 268f45cc-cd09-0410-ab3c-d52691b4dbfc