[Modern Media Controls] iOS: long press on play button shows the Copy menu https://bugs.webkit.org/show_bug.cgi?id=169355 <rdar://problem/30928998>
Patch by Antoine Quint <graouts@apple.com> on 2017-03-09
Reviewed by Dean Jackson.

Source/WebCore:

Disable built-in UI behavior that shows a magnifier or a copy callout when pressing
on an element for a long time. These do not make sense for media controls.

Test: media/modern-media-controls/audio/audio-controls-styles.html

* Modules/modern-media-controls/controls/media-controls.css:
(:host):

LayoutTests:

Add a test to check that the expected properties are set as they should to disable
built-in UI behavior that shows a magnifier or a copy callout when pressing on an
element for a long time.

* media/modern-media-controls/audio/audio-controls-styles-expected.txt: Added.
* media/modern-media-controls/audio/audio-controls-styles.html: Added.
* platform/mac/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213670 268f45cc-cd09-0410-ab3c-d52691b4dbfc