[WK2][CMake] Remove the intermediate NetworkProcess_COMMON_SOURCES variable https://bugs.webkit.org/show_bug.cgi?id=169524
Reviewed by Michael Catanzaro.

The NetworkProcess-specific source files listed in the
NetworkProcess_COMMON_SOURCES variable definition should be directly
inlined into the list with which the WebKit2_SOURCE variable is defined.
The NetworkProcess_COMMON_SOURCES variable is appended to the
WebKit2_SOURCES variable anyway, and is not used anywhere else. The
separate variable was previously use to support configurations that
allowed disabling NetworkProcess.

* CMakeLists.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213787 268f45cc-cd09-0410-ab3c-d52691b4dbfc