Web Inspector: add context menu item to log content of WebSocket frame https://bugs.webkit.org/show_bug.cgi?id=169945
Reviewed by Joseph Pecoraro.

* Localizations/en.lproj/localizedStrings.js:

* UserInterface/Views/WebSocketContentView.js:
(WebInspector.WebSocketContentView):
(WebInspector.WebSocketContentView.prototype.addFrame):
(WebInspector.WebSocketContentView.prototype._addRow):
Rework the way CSS classes are applied so the isOutgoing and isText flags can also be passed
to the WebSocketDataGridNode.  This is necessary because a non-text frame shouldn't be able
to be logged to the console.

* UserInterface/Views/WebSocketDataGridNode.js:
(WebInspector.WebSocketDataGridNode.prototype.appendContextMenuItems):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214371 268f45cc-cd09-0410-ab3c-d52691b4dbfc