[Modern Media Controls] Controls bar may disappear while captions menu is visible (redux) https://bugs.webkit.org/show_bug.cgi?id=170239 <rdar://problem/31320685>
Patch by Antoine Quint <graouts@apple.com> on 2017-03-29
Reviewed by Dean Jackson.

Source/WebCore:

We did some work in webkit.org/b/168751 to prevent the controls bar from disappearing while the
captions menu is visible. But there were two cases where the behavior was not as intended:

    1. the controls bar would hide upon exiting the video.
    2. clicking on the controls bar while the caption panel is up would hide the controls bar
       as well as the captions panel.

Instead of determining that the "userInteractionEnabled" property being set to "false" is indicative
of secondary UI, such as the tracks panel, being attached to the controls bar, we now have an
explicit property to specify this. Now, when "hasSecondaryUIAttached" is "true", we prevent the
controls bar from fading when exiting the media.

Additionally, when dismissing the tracks panel, we check whether a mouse event lead to this and check
if the mouse was positioned over the controls bar. If that is the case, we no longer hide the controls
bar and only dismiss the tracks panel.

Test: media/modern-media-controls/tracks-panel/tracks-panel-controls-bar-remains-visible-after-clicking-over-it.html

* Modules/modern-media-controls/controls/controls-bar.js:
(ControlsBar.prototype.set userInteractionEnabled):
(ControlsBar.prototype.handleEvent):
(ControlsBar.prototype._autoHideTimerFired):
* Modules/modern-media-controls/controls/macos-media-controls.js:
(MacOSMediaControls.prototype.showTracksPanel):
(MacOSMediaControls.prototype.hideTracksPanel):

LayoutTests:

We extend an existing test to cover the case where we would hide the controls bar while the tracks panel
is up by mousing out of the media. We also add a new test where we check that clicking over the controls
bar to dismiss the tracks panel does not hide the controls bar. Finally, we make a small change to an
existing test to listen to "transitionend" event on the tracks panel specifically since with this patch
the controls bar would disappear and trigger a similar event when disappearing as the tracks panel is dismissed.

* media/modern-media-controls/tracks-panel/tracks-panel-controls-bar-remains-visible-after-clicking-over-it-expected.txt: Added.
* media/modern-media-controls/tracks-panel/tracks-panel-controls-bar-remains-visible-after-clicking-over-it.html: Added.
* media/modern-media-controls/tracks-panel/tracks-panel-prevent-controls-bar-from-fading-expected.txt:
* media/modern-media-controls/tracks-panel/tracks-panel-prevent-controls-bar-from-fading.html:
* media/modern-media-controls/tracks-support/tracks-support-click-track-in-panel.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214545 268f45cc-cd09-0410-ab3c-d52691b4dbfc