[Modern Media Controls] Invalid placard icon is not visible for short video https://bugs.webkit.org/show_bug.cgi?id=170277 <rdar://problem/31327955>
Patch by Antoine Quint <graouts@apple.com> on 2017-03-30
Reviewed by Eric Carlson.

Source/WebCore:

Make the minimum dimension needed to display a placard icon configurable and let the
invalid placard use a 50pt dimension. We also make a drive-by fix to reset "text-align"
for media controls as having a "text-align: center" inherited value would mess up
positioning of the placard icon and make it flush to the right of the placard.

Test: media/modern-media-controls/invalid-placard/invalid-placard-constrained-metrics.html

* Modules/modern-media-controls/controls/invalid-placard.js:
(InvalidPlacard):
* Modules/modern-media-controls/controls/media-controls.css:
(.media-controls):
* Modules/modern-media-controls/controls/placard.js:
(Placard.prototype.layout):

LayoutTests:

Add a new test to check that we correctly hide the invalid icon only when
there is no space left to display it.

* media/modern-media-controls/invalid-placard/invalid-placard-constrained-metrics-expected.txt: Added.
* media/modern-media-controls/invalid-placard/invalid-placard-constrained-metrics.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214611 268f45cc-cd09-0410-ab3c-d52691b4dbfc