Web Inspector: Assertion failure in InspectorStyleProperty::setRawTextFromStyleDeclaration https://bugs.webkit.org/show_bug.cgi?id=170279 <rdar://problem/30200492>
Reviewed by David Hyatt.

Source/WebCore:

The SourceRange for a CSSPropertySourceData should be relative to the start
of the declaration body, not the start of the StyleSheetHandler’s parsed
text. This only affected the ranges of unparsed (parsedOK == false) properties
lacking a trailing semi-colon.

This patch doesn't change the behavior of InspectorStyleSheet other than
silencing an irksome assert, as String::substring does a safety check on
the passed in length.

* inspector/InspectorStyleSheet.cpp:
(WebCore::fixUnparsedProperties):

LayoutTests:

* inspector/css/getAllStyleSheets.html:
Add stylesheet with an "unparsed" property having no final semicolon.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214634 268f45cc-cd09-0410-ab3c-d52691b4dbfc