YouTube sometimes does not respect "user gesture" restriction for videos. https://bugs.webkit.org/show_bug.cgi?id=170297
I discovered a code path that does not honor the "user gesture" requirement and playback is able to begin
even though we have a restriction in place. When using Media Source Extensions, which YouTube does, we transition
from a "Have Metadata" to a "Future Data" state that causes playback to begin, however, we never check
if we have a playback restriction in place.

Reviewed by Eric Carlson.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::setReadyState):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214639 268f45cc-cd09-0410-ab3c-d52691b4dbfc