REGRESSION (r206744): CSS background-image in style attribute ignored when using createHTMLDocument method of DOM parsing https://bugs.webkit.org/show_bug.cgi?id=170285 <rdar://problem/31378543>
Reviewed by Andy Estes.

Source/WebCore:

r206744 caused up to stop trying to resolve relative URLs when trying to load an image
referred to by CSS. We already try to resolve the relative URL when parsing the CSS
property so this will usually work fine. However, in the case when the CSS property
is parsed in detached document and then moved to another document, we will not have
the complete URL.

Test: fast/images/background-image-relative-url-changes-document.html

* css/CSSImageValue.cpp:
(WebCore::CSSImageValue::loadImage):

LayoutTests:

Add layout test coverage.

* fast/images/background-image-relative-url-changes-document-expected.html: Added.
* fast/images/background-image-relative-url-changes-document.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214842 268f45cc-cd09-0410-ab3c-d52691b4dbfc