[CMake][Windows] WebKitGUID.lib should be built with the release CRT https://bugs.webkit.org/show_bug.cgi?id=170593
Patch by Fujii Hironori <Hironori.Fujii@sony.com> on 2017-04-07
Reviewed by Brent Fulgham.

Source/cmake/OptionsWin.cmake has a trick to force using the
release CRT even for debug build configuration. But, this works
only for C++, but for C. As the result, WebKitGUID.lib is built
with the debug CRT in debug build configuration because it has C
source files.

* Source/cmake/OptionsWin.cmake: Process CMAKE_C_FLAGS_* as well as CMAKE_CXX_FLAGS_*.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215097 268f45cc-cd09-0410-ab3c-d52691b4dbfc