Unprefix CSS cursor values grab and grabbing https://bugs.webkit.org/show_bug.cgi?id=170543
Reviewed by Jon Lee.

Source/WebCore:

Add support for unprefixed "grab" and "grabbing" values for cursor (retaining
support for the prefixed values) which are now in <https://drafts.csswg.org/css-ui-3/#cursor>

Canonicalize the order of the values based on the order in the spec in the enums
and switch statements.

Tested by fast/css/cursor-parsing.html

* css/CSSPrimitiveValueMappings.h:
(WebCore::CSSPrimitiveValue::CSSPrimitiveValue):
(WebCore::CSSPrimitiveValue::operator ECursor):
* css/CSSProperties.json:
* css/CSSValueKeywords.in:
* page/EventHandler.cpp:
(WebCore::EventHandler::selectCursor):
* rendering/style/RenderStyleConstants.h:

Source/WebInspectorUI:

Sort the cursor autosuggestions to match the order in the spec, and add the unprefixed
grab and grabbing values.

* UserInterface/Models/CSSKeywordCompletions.js:

LayoutTests:

Enhanced the cursor test to check round-tripping of all values.

* fast/css/cursor-parsing-expected.txt:
* fast/css/cursor-parsing.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215146 268f45cc-cd09-0410-ab3c-d52691b4dbfc