WebAssembly: There is a short window of time where a CodeBlock could be destroyed before all of its async compilation callbacks are called https://bugs.webkit.org/show_bug.cgi?id=170641
Reviewed by Keith Miller.

There is an unlikely race when a CodeBlock compilation fails,
the module compiles a new CodeBlock for that memory mode, all while
the CodeBlock is notifying its callbacks that it has finished.
There is a chance that the Module could deref its failed CodeBlock
at that point, destroying it, before the callbacks were able to
grab a Ref to the CodeBlock. This patch fixes the race by having the
callbacks ref the CodeBlock.

This patch also has the Plan clear out all of its callbacks
once it gets completed. This adds an extra defense to anybody
that grabs refs to the Plan in the callback.

* wasm/WasmCodeBlock.cpp:
(JSC::Wasm::CodeBlock::CodeBlock):
(JSC::Wasm::CodeBlock::compileAsync):
* wasm/WasmPlan.cpp:
(JSC::Wasm::Plan::complete):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215368 268f45cc-cd09-0410-ab3c-d52691b4dbfc