[WK2] Cannot select text on nytimes.com when the selection granularity is WKSelectionGranularityCharacter https://bugs.webkit.org/show_bug.cgi?id=170968 <rdar://problem/31692560>
Reviewed by Dan Bernstein.

Source/WebKit2:

Makes a small adjustment to textInteractionGesture:shouldBeginAtPoint:. When determining whether to allow the
text interaction assistant to recognize at a given point, instead of depending on whether or not the hit node is
the same as the assisted node, only enforce this restriction when editing an assisted node. Otherwise, default
to allowing the selection gesture.

Note that character granularity selection was working on most pages before, due to the fact that
nodeAtPositionIsAssistedNode was true in many cases when there is no assisted node at all. This is because, in
WebPage.mm, we compute the hit-tested node responding to click events to be null, and then set
nodeAtPositionIsAssistedNode to be equal to hitNode == m_assistedNode, which ends up being true because both of
these values are null.

This allowed text selection to work in the simple case when selection granularity character is used, but not
when the node containing the selected point actually does respond to click events, since the comparison returns
false.

New layout test: LayoutTests/editing/selection/character-granularity-select-text-with-click-handler.html.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView textInteractionGesture:shouldBeginAtPoint:]):

LayoutTests:

Adds a new layout test checking that text within a node with a click handler can be selected when using
character selection granularity.

* editing/selection/character-granularity-select-text-with-click-handler-expected.txt: Added.
* editing/selection/character-granularity-select-text-with-click-handler.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215484 268f45cc-cd09-0410-ab3c-d52691b4dbfc