Fix assertions in webProcessProxyFromConnection https://bugs.webkit.org/show_bug.cgi?id=171025 <rdar://problem/31184073>
Patch by Alex Christensen <achristensen@webkit.org> on 2017-04-20
Reviewed by Brady Eidson.

* UIProcess/WebProcessPool.cpp:
(WebKit::webProcessProxyFromConnection):
If we have a Connection, then we have a valid process attached to that connection,
unless it's a Connection to a different type of child process.
Calling ChildProcessProxy::connection on other web processes that are in State::Launching,
then we get an assertion.  Luckily, ChildProcessProxy::hasConnection was introduced in r210861
for this reason.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215567 268f45cc-cd09-0410-ab3c-d52691b4dbfc