Improve wavy underline rendering https://bugs.webkit.org/show_bug.cgi?id=171281
Reviewed by Sam Weinig.

Change wavy underline rendering to use font size instead of stroke
thickness as a way to tune it. Cut the extra offset from 2 to 1.
Change the parameters to make the line flatter and to make the proportion
of the wave that renders underneath a glyph the same.

* rendering/TextDecorationPainter.cpp:
(WebCore::strokeWavyTextDecoration):
(WebCore::TextDecorationPainter::paintTextDecoration):
* style/InlineTextBoxStyle.cpp:
(WebCore::getWavyStrokeParameters):
(WebCore::visualOverflowForDecorations):
* style/InlineTextBoxStyle.h:
(WebCore::wavyOffsetFromDecoration):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215802 268f45cc-cd09-0410-ab3c-d52691b4dbfc