Repeated layouts in Mail due to viewport units being used with auto-sizing https://bugs.webkit.org/show_bug.cgi?id=171371 <rdar://problem/28780084>
Reviewed by Zalan Bujtas.

Source/WebCore:

Test: css3/viewport-percentage-lengths/vh-auto-size.html

Auto-sizing code would adjust the size of the view in the beginning of layout(). This would
end up invalidating style for elements that use vh units and we would perform main layout
with unclean style. This would result in endless layout loops and hit assert on debug.

* page/FrameView.cpp:
(WebCore::FrameView::availableContentSizeChanged):

    Ensure we heve clean style after resize if we are in pre-layout.

LayoutTests:

* css3/viewport-percentage-lengths/vh-auto-size-expected.html: Added.
* css3/viewport-percentage-lengths/vh-auto-size.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215874 268f45cc-cd09-0410-ab3c-d52691b4dbfc