On-screen panel for candidate bar is in the wrong place when the caret is at the start of a paragraph https://bugs.webkit.org/show_bug.cgi?id=171453 -and corresponding- rdar://problem/29779764
Reviewed by Tim Horton.

Source/WebKit/mac:

This bug only happens when the caret is at the beginning of the paragraph, and it 
is because Range::absoluteTextQuads() returns no quads in this case. I think it 
might be correct that absoluteTextQuads() returns no quads in this case, so this 
patch fixes the bug at the call site by using absoluteCaretBounds() in this case.

* WebCoreSupport/WebEditorClient.mm:
(WebEditorClient::handleRequestedCandidates):

Source/WebKit2:

This bug only happens when the caret is at the beginning of the paragraph, and it 
is because Range::absoluteTextQuads() returns no quads in this case. I think it 
might be correct that absoluteTextQuads() returns no quads in this case, so this 
patch fixes the bug at the call site by using absoluteCaretBounds() in this case.

SPI for testing.
* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _candidateRect]):
* UIProcess/API/Cocoa/WKWebViewPrivate.h:

Bug fix.
* WebProcess/WebPage/mac/WebPageMac.mm:
(WebKit::WebPage::platformEditorState):

Tools:

This test hard-codes the y-coordinate for the expected candidateRect. 
* TestWebKitAPI/Tests/WebKit2Cocoa/WKWebViewCandidateTests.mm:
(TEST):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216018 268f45cc-cd09-0410-ab3c-d52691b4dbfc