check-webkit-style exits with an error while parsing Source/WebKit2/Scripts/webkit/ directory: KeyError: 'st' <https://bugs.webkit.org/show_bug.cgi?id=171510>
Reviewed by Daniel Bates.

The problem was that 'import parser' was ambiguous since there
is a built-in parser module that comes with Python.  Changing
'import parser' to 'from webkit import parser' fixed the error
in check-webkit-style (and pylint) by forcing it to load the
local 'parser' module, but then running messages_unittest.py
directly would fail because .../Source/WebKit2/Scripts was not
in the module search path by default.  This is fixed by using
sys.path.append() to add .../Source/WebKit2/Scripts to the
module search path based on the path to messages_unittest.py.

As an added bonus, this patch also fixes all webkit-style issues
in Source/WebKit2/Scripts/webkit/*.py files.

* Scripts/webkit/messages.py:
- Separate local module imports from built-in modules for
  consistency.
- Fix whitespace issues for check-webkit-style, and remove stray
  semi-colon.
* Scripts/webkit/messages_unittest.py:
- Update sys.path to find the 'webkit' module based on the path
  to messages_unittest.py.  This makes 'from webkit import ...'
  statements work when running messages_unittest.py directly.
- Change 'messages' and 'parser' modules to use
  'from webkit import ...' format for consistency.
- Fix whitespace issues for check-webkit-style.
* Scripts/webkit/model.py:
- Fix whitespace issue for check-webkit-style.
* Scripts/webkit/parser.py:
- Fix whitespace issue for check-webkit-style.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216181 268f45cc-cd09-0410-ab3c-d52691b4dbfc