[CG] To decode an image frame asynchronously, pass the native size of the image to CGImageSourceCreateThumbnailAtIndex() if it's smaller than the sizeForDrawing https://bugs.webkit.org/show_bug.cgi?id=170864 Source/WebCore:
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2017-05-08
Reviewed by Simon Fraser.

Tests: Existing tests were modified to test this patch.

The maxPixelSize was omitted when calling CGImageSourceCreateThumbnailAtIndex()
for the nativeSize image or when areaOf(nativeSize) < areaOf(sizeForDrawing).
The assumption was if we don't pass maxPixelSize, CG will create an image with the
nativeSize regardless how big this size is. It turns out this is wrong. CG has an
optimization to return a scaled down image if areaOf(nativeSize) is greater than
some maximum value.

This is not what we want for asynchronously image decoding. We want the decoded
frame to have the size we ask for, not some scaled down frame. The fix is to pass
always maxPixelSize to CGImageSourceCreateThumbnailAtIndex(). maxPixelSize will be
equal to maxDimension(sizeForDrawing) if sizeForDrawing is valid it is less than
maxDimension(nativeSize). Otherwise it will be equal to maxDimension(nativeSize).

* platform/graphics/DecodingOptions.h:
* platform/graphics/cg/ImageDecoderCG.cpp:
(WebCore::appendImageSourceOption):
(WebCore::appendImageSourceOptions):
(WebCore::imageSourceAsyncOptions):
(WebCore::ImageDecoder::createFrameImageAtIndex):

LayoutTests:

Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2017-05-08
Reviewed by Simon Fraser.

* fast/images/async-image-background-image-repeated-expected.html:
* fast/images/async-image-background-image-repeated.html:
* fast/images/resources/sprite-sheet-red-green-blue.png:
* fast/images/sprite-sheet-image-draw-expected.html:
* fast/images/sprite-sheet-image-draw.html:
Change the sprite image to be 200x33100 instead of 200x2100 and use it in drawing.
If the maxPixelSize is not passed to CGImageSourceCreateThumbnailAtIndex(), the
decoded farme of this image will be 30x5000 pixels (15% of the original size).

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216450 268f45cc-cd09-0410-ab3c-d52691b4dbfc