Simple line layout: Leading whitespace followed by a <br> produces an extra linebreak. https://bugs.webkit.org/show_bug.cgi?id=172076
Reviewed by Antti Koivisto.

Source/WebCore:

When the collapsed whitespace does not fit the line, we need to push it to the next line
so that we can decide whether any soft/hard linebreak should be skipped (to avoid double line breaks) or not.

Test: fast/text/simple-line-layout-leading-whitespace-with-soft-hard-linebreak.html

* rendering/SimpleLineLayout.cpp:
(WebCore::SimpleLineLayout::consumeLineBreakIfNeeded): special handling <br>
(WebCore::SimpleLineLayout::firstFragment): Now we need to deal with leading collapsed whitespace.
(WebCore::SimpleLineLayout::createLineRuns): We need to push even the collapsed whitespace to the next line.

LayoutTests:

* fast/text/simple-line-layout-leading-whitespace-with-soft-hard-linebreak-expected.html: Added.
* fast/text/simple-line-layout-leading-whitespace-with-soft-hard-linebreak.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216861 268f45cc-cd09-0410-ab3c-d52691b4dbfc