[WPE] GLContextEGL::createWPEContext() should fall back to pbuffer-based contexts when offscreen target provides no native window https://bugs.webkit.org/show_bug.cgi?id=172162
Reviewed by Carlos Garcia Campos.

When creating an offscreen GLContext, the underlying implementation might
provide a mock native window that's to be used as the window target upon
which a window-based GLContext should be created. But we should also support
falling back to pbuffer-based GLContexts when the underlying implementation
can't provide such mock targets.

* platform/graphics/egl/GLContextEGLWPE.cpp:
(WebCore::GLContextEGL::createWPEContext):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216930 268f45cc-cd09-0410-ab3c-d52691b4dbfc