Setting URL.search to '' results in a stringified URL ending in '?' https://bugs.webkit.org/show_bug.cgi?id=162345 <rdar://problem/31800441>
Reviewed by Alex Christensen.

LayoutTests/imported/w3c:

Rebaseline web-platform-test now that more checks are passing.

* web-platform-tests/url/url-setters-expected.txt:

Source/WebCore:

As per the specification for the URL.search setter [1], if the given value is
the empty string, then we should set the URL's query to null. We would
previously set the URL's query to the empty string in this case. This aligns
our behavior with Firefox and Chrome.

[1] https://url.spec.whatwg.org/#dom-url-search

No new tests, updated existing tests.

* html/URLUtils.h:
(WebCore::URLUtils<T>::setSearch):

LayoutTests:

* fast/dom/DOMURL/set-href-attribute-search-expected.txt:
* fast/dom/DOMURL/set-href-attribute-search.html:
* fast/dom/HTMLAnchorElement/script-tests/set-href-attribute-search.js:
* fast/dom/HTMLAnchorElement/set-href-attribute-search-expected.txt:
Update / rebaseline existing tests to reflect behavior change.

* platform/ios-wk1/TestExpectations:
* platform/mac-wk1/TestExpectations:
Unskip url-setters.html test as it seems to be passing for me on WK1.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217004 268f45cc-cd09-0410-ab3c-d52691b4dbfc