Avoid moving to the second frame of an animated image before the first frame has finished decoding https://bugs.webkit.org/show_bug.cgi?id=172461
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2017-05-22
Reviewed by Simon Fraser.

Sometimes the frameCount of an animated image gets changed when more data
is received. The problem is we may start decoding the image as if it were
a static large image and the decoding of this first frame finishes after
the image starts animating.

* platform/graphics/BitmapImage.cpp:
(WebCore::BitmapImage::draw):
(WebCore::BitmapImage::imageFrameAvailableAtIndex):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217262 268f45cc-cd09-0410-ab3c-d52691b4dbfc