Stop consulting -viewsNeedDisplay when doing layer flushing in WebKit1 https://bugs.webkit.org/show_bug.cgi?id=172522 rdar://problem/31071812
Source/WebKit/mac:

Reviewed by Tim Horton.

There's no need to consult -[NSWindow viewsNeedDisplay] on newer OSes. It was
added in r58623, and modified in r61204 for Carbon apps, but with more recent AppKit
versions this synchronization is no longer required.

* WebView/WebView.mm:
(LayerFlushController::flushLayers):

Tools:

Reviewed by Sam Weinig.

Remove NSWindowStyleMaskFullSizeContentView which is no longer referenced in code.
We set this in the .xib file now.

* MiniBrowser/mac/MiniBrowser_Prefix.pch:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217303 268f45cc-cd09-0410-ab3c-d52691b4dbfc