REGRESSION(r216882): No image decoding is needed if the BitmapImage is created with a NativeImage https://bugs.webkit.org/show_bug.cgi?id=172670
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2017-05-30
Reviewed by Tim Horton.

Source/WebCore:

Test: fast/images/large-image-webkit-canvas.html

Check whether the BitmapImage has created an ImageDecoder before trying
to decode its image frame.

* platform/graphics/ImageSource.cpp:
(WebCore::ImageSource::shouldUseAsyncDecoding):

LayoutTests:

Use the css property -webkit-canvas to force capturing the canvas context
in a NativeImage. This NativeImage will be used to create a BitmapImage
which will be later used as the background-image of an element.

* fast/images/large-image-webkit-canvas-expected.html: Added.
* fast/images/large-image-webkit-canvas.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217567 268f45cc-cd09-0410-ab3c-d52691b4dbfc