Web Inspector: images dragged from Inspector to Desktop are named "Unknown.png" https://bugs.webkit.org/show_bug.cgi?id=141515 <rdar://problem/9251308>
Reviewed by Wenson Hsieh.

Source/WebCore:

* html/HTMLAttributeNames.in:

Source/WebInspectorUI:

* UserInterface/Views/ImageResourceContentView.js:
(WebInspector.ImageResourceContentView.prototype.contentAvailable):
Set a non-standard "filename" attribute to provide a suggested filename
for this <img> containing data: or blob: content.

Source/WebKit2:

In Web Inspector these images are <img src="..."> with data: or blob:
content. In these cases we have the image data but the URL does not
provide a useful name. In fact the name "Unknown.png" comes from
-[NSURLResponse suggestedFilename] for this URL / content.

To start, improve this situation for drags in Web Inspector pages.
If the image element being dragged has a filename attribute use that
as the suggested filename for the drag.

* WebProcess/WebCoreSupport/mac/WebDragClientMac.mm:
(WebKit::WebDragClient::declareAndWriteDragImage):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217584 268f45cc-cd09-0410-ab3c-d52691b4dbfc